split core/commands/dag into individual files for different subcommands
